# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlphaTrade Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List recommendations (latest first)
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by date (YYYY-MM-DD)
        - name: month
          in: query
          schema:
            type: string
          description: Filter by month (YYYY-MM)
        - name: tradeType
          in: query
          schema:
            $ref: "#/components/schemas/TradeType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RecommendationStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationList"
        "403":
          description: Subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createRecommendation
      tags: [recommendations]
      summary: Create a stock recommendation (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRequest"
      responses:
        "201":
          description: Recommendation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recommendations/today:
    get:
      operationId: getTodayRecommendations
      tags: [recommendations]
      summary: Get today's recommendations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today's recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "403":
          description: Subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recommendations/{id}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      summary: Get a single recommendation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateRecommendation
      tags: [recommendations]
      summary: Update a recommendation (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecommendationRequest"
      responses:
        "200":
          description: Updated recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"

    delete:
      operationId: deleteRecommendation
      tags: [recommendations]
      summary: Delete a recommendation (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /recommendations/{id}/pnl:
    put:
      operationId: updateRecommendationPnl
      tags: [pnl]
      summary: Update evening P&L for a recommendation (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePnlRequest"
      responses:
        "200":
          description: Updated recommendation with P&L
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create Razorpay order for subscription
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Razorpay order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify Razorpay payment and activate subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified, subscription activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
  /payments/history:
    get:
      operationId: getPaymentHistory
      tags: [payments]
      summary: Get user payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    TradeType:
      type: string
      enum: [intraday, swing, positional]
    RiskLevel:
      type: string
      enum: [low, medium, high]
    RecommendationStatus:
      type: string
      enum: [active, target_hit, stop_loss_hit, hold, partial_profit, closed]
    SubscriptionType:
      type: string
      enum: [trial, premium, expired]
...
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        stockName:
          type: string
        nseSymbol:
          type: string
        buyPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        tradeType:
          $ref: "#/components/schemas/TradeType"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/RecommendationStatus"
        exitPrice:
          type: number
          nullable: true
        pnlPercent:
          type: number
          nullable: true
        pnlAbsolute:
          type: number
          nullable: true
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, stockName, nseSymbol, buyPrice, targetPrice, stopLoss, tradeType, riskLevel, status, date, createdAt, updatedAt]
    RecommendationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    CreateRecommendationRequest:
      type: object
      required: [stockName, nseSymbol, buyPrice, targetPrice, stopLoss, tradeType, riskLevel]
      properties:
        stockName:
          type: string
        nseSymbol:
          type: string
        buyPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        tradeType:
          $ref: "#/components/schemas/TradeType"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        notes:
          type: string
        date:
          type: string
          format: date
    UpdateRecommendationRequest:
      type: object
      properties:
        stockName:
          type: string
        nseSymbol:
          type: string
        buyPrice:
          type: number
        targetPrice:
          type: number
        stopLoss:
          type: number
        tradeType:
          $ref: "#/components/schemas/TradeType"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        notes:
          type: string
        date:
          type: string
          format: date
    UpdatePnlRequest:
      type: object
      required: [exitPrice, status]
      properties:
        exitPrice:
          type: number
        status:
          $ref: "#/components/schemas/RecommendationStatus"
        notes:
          type: string
...
    SubscriptionStatus:
      type: object
      properties:
        subscriptionType:
          $ref: "#/components/schemas/SubscriptionType"
        trialStartDate:
          type: string
          format: date-time
          nullable: true
        trialExpiryDate:
          type: string
          format: date-time
          nullable: true
        premiumExpiryDate:
          type: string
          format: date-time
          nullable: true
        daysRemaining:
          type: integer
          nullable: true
        isActive:
          type: boolean
      required: [subscriptionType, isActive]
    PaymentOrder:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
        keyId:
          type: string
      required: [orderId, amount, currency, keyId]
    VerifyPaymentRequest:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
    Payment:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        razorpayOrderId:
          type: string
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, amount, currency, status, createdAt]
...