# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationStatus } from './recommendationStatus';
import type { RiskLevel } from './riskLevel';
import type { TradeType } from './tradeType';

export interface Recommendation {
  id: number;
  stockName: string;
  nseSymbol: string;
  buyPrice: number;
  targetPrice: number;
  stopLoss: number;
  tradeType: TradeType;
  riskLevel: RiskLevel;
  notes?: string | null;
  status: RecommendationStatus;
  exitPrice?: number | null;
  pnlPercent?: number | null;
  pnlAbsolute?: number | null;
  date: Date;
  createdAt: Date;
  updatedAt: Date;
}