# lib/api-zod/src/generated/types/recommendationStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export type RecommendationStatus = typeof RecommendationStatus[keyof typeof RecommendationStatus];


export const RecommendationStatus = {
  active: 'active',
  target_hit: 'target_hit',
  stop_loss_hit: 'stop_loss_hit',
  hold: 'hold',
  partial_profit: 'partial_profit',
  closed: 'closed',
} as const;