# lib/api-zod/src/generated/types/updatePnlRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaTrade Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationStatus } from './recommendationStatus';

export interface UpdatePnlRequest {
  exitPrice: number;
  status: RecommendationStatus;
  notes?: string;
}