# lib/db/src/schema/recommendations.ts
import { pgTable, text, serial, timestamp, numeric, date } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const recommendationsTable = pgTable("recommendations", {
  id: serial("id").primaryKey(),
  stockName: text("stock_name").notNull(),
  nseSymbol: text("nse_symbol").notNull(),
  buyPrice: numeric("buy_price", { precision: 10, scale: 2 }).notNull(),
  targetPrice: numeric("target_price", { precision: 10, scale: 2 }).notNull(),
  stopLoss: numeric("stop_loss", { precision: 10, scale: 2 }).notNull(),
  tradeType: text("trade_type").notNull(),
  riskLevel: text("risk_level").notNull().default("medium"),
  notes: text("notes"),
  status: text("status").notNull().default("active"),
  exitPrice: numeric("exit_price", { precision: 10, scale: 2 }),
  pnlPercent: numeric("pnl_percent", { precision: 10, scale: 4 }),
  pnlAbsolute: numeric("pnl_absolute", { precision: 10, scale: 2 }),
  date: date("date").notNull(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertRecommendationSchema = createInsertSchema(recommendationsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertRecommendation = z.infer<typeof insertRecommendationSchema>;
export type Recommendation = typeof recommendationsTable.$inferSelect;